Refactor CLI server commands
